+++ tf_encrypted/keras/layers/__init__.py
@@ -3,14 +3,13 @@ from __future__ import absolute_import
 
 from tf_encrypted.keras.engine.input_layer import Input
 from tf_encrypted.keras.layers.activation import Activation
-from tf_encrypted.keras.layers.convolutional import Conv2D
+from tf_encrypted.keras.layers.convolutional import Conv2D, DepthwiseConv2D
 from tf_encrypted.keras.layers.dense import Dense
 from tf_encrypted.keras.layers.flatten import Flatten
 from tf_encrypted.keras.layers.pooling import AveragePooling2D, MaxPooling2D
 from tf_encrypted.keras.layers.relu import ReLU
 from tf_encrypted.keras.layers.normalization import BatchNormalization
 from tf_encrypted.keras.layers.core import Reshape
-from tf_encrypted.keras.layers.depthwise_conv import DepthwiseConv2D
 
 
 __all__ = [
